SPARK-29083 Applied the same PrefetchingIterator for Python
